refactor: add CreateSpecForm component with form fields and validation
This commit introduces the `CreateSpecForm` component which includes form fields for "Summary", "Technical Constraints", and "Open Questions". Each field is accompanied by appropriate labels, placeholder text, and validation logic. The form also integrates with a text area panel component for consistent styling and includes support for accessibility attributes like `aria-required` and `aria-invalid`. The component is designed to be reusable within the create specification view, supporting features such as character counting, error message display, and integration with form submission handlers.
